+++ www/admin/account-preferences-plugin.php
@@ -46,7 +46,8 @@ $enabled = $GLOBALS['_MAX']['CONF']['pluginGroupComponents'][$group];
 $disabled = ((!$enabled) && (OA_Permission::getAccountType() != OA_ACCOUNT_ADMIN));
 
 // If the settings page is a submission, deal with the form data
-if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
+$oSettings = new OA_Admin_Settings();
+if ($oSettings->isFormSubmitted()) {
     // Prepare an array of the HTML elements to process, and which
     // of the preferences are checkboxes
     $aElements = [];
@@ -74,7 +75,7 @@ if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
     }
 
     if ($valid) {
-        // Create a new settings object, and save the settings!
+        // Save the settings!
         $result = OA_Preferences::processPreferencesFromForm($aElements, $aCheckboxes);
         if ($result) {
             // Queue confirmation message