+++ www/admin/account-preferences-tracker.php
@@ -35,7 +35,8 @@ $prefSection = "tracker";
 $aErrormessage = [];
 
 // If the settings page is a submission, deal with the form data
-if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
+$oSettings = new OA_Admin_Settings();
+if ($oSettings->isFormSubmitted()) {
     // Prepare an array of the HTML elements to process, and which
     // of the preferences are checkboxes
     $aElements = [];