+++ www/admin/account-settings-banner-storage.php
@@ -32,7 +32,8 @@ $prefSection = "banner-storage";
 $aErrormessage = [];
 
 // If the settings page is a submission, deal with the form data
-if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
+$oSettings = new OA_Admin_Settings();
+if ($oSettings->isFormSubmitted()) {
     // Prepare an array of the HTML elements to process, and the
     // location to save the values in the settings configuration
     // file
@@ -148,7 +149,7 @@ if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
         }
     }
     if (empty($aErrormessage)) {
-        // Create a new settings object, and save the settings!
+        // Save the settings!
         $oSettings = new OA_Admin_Settings();
         $result = $oSettings->processSettingsFromForm($aElements);
         if ($result) {