+++ www/admin/account-settings-database.php
@@ -33,7 +33,8 @@ $aErrormessage = [];
 
 // FORM IS NOW DISABLED
 // If the settings page is a submission, deal with the form data
-/*if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
+/*$oSettings = new OA_Admin_Settings();
+if ($oSettings->isFormSubmitted()) {
     // Prepare an array of the HTML elements to process, and the
     // location to save the values in the settings configuration
     // file
@@ -108,7 +109,7 @@ $aErrormessage = [];
     // if we managed to connect using both methods, go ahead and save the db connection details
     if ($connected)
     {
-        // Create a new settings object, and save the settings!
+        // Save the settings!
         $oSettings = new OA_Admin_Settings();
         $result = $oSettings->processSettingsFromForm($aElements);
         if ($result) {