+++ www/admin/account-settings-debug.php
@@ -32,7 +32,8 @@ $prefSection = "debug";
 $aErrormessage = [];
 
 // If the settings page is a submission, deal with the form data
-if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
+$oSettings = new OA_Admin_Settings();
+if ($oSettings->isFormSubmitted()) {
     // Prepare an array of the HTML elements to process, and the
     // location to save the values in the settings configuration
     // file
@@ -78,8 +79,7 @@ if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
         'log_paramsUsername' => ['log' => 'paramsUsername'],
         'log_paramsPassword' => ['log' => 'paramsPassword'],
     ];
-    // Create a new settings object, and save the settings!
-    $oSettings = new OA_Admin_Settings();
+    // Save the settings!
     $result = $oSettings->processSettingsFromForm($aElements);
     if ($result) {
         // Queue confirmation message