+++ www/admin/account-settings-maintenance.php
@@ -33,7 +33,8 @@ $prefSection = "maintenance";
 $aErrormessage = [];
 
 // If the settings page is a submission, deal with the form data
-if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
+$oSettings = new OA_Admin_Settings();
+if ($oSettings->isFormSubmitted()) {
     // Prepare an array of the HTML elements to process, and the
     // location to save the values in the settings configuration
     // file
@@ -53,8 +54,7 @@ if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
         ],
         'priority_intentionalOverdelivery' => ['priority' => 'intentionalOverdelivery'],
     ];
-    // Create a new settings object, and save the settings!
-    $oSettings = new OA_Admin_Settings();
+    // Save the settings!
     $result = $oSettings->processSettingsFromForm($aElements);
     if ($result) {
         // Queue confirmation message