+++ www/admin/plugin-settings.php
@@ -47,7 +47,8 @@ $oManager = new OX_Plugin_ComponentGroupManager();
 $aComponentSettings = $oManager->getComponentGroupSettings($group, true);
 
 // If the settings page is a submission, deal with the form data
-if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
+$oSettings = new OA_Admin_Settings();
+if ($oSettings->isFormSubmitted()) {
     // Prepare an array of the HTML elements to process, and the
     // location to save the values in the settings configuration
     // file
@@ -81,8 +82,7 @@ if (isset($_POST['submitok']) && $_POST['submitok'] == 'true') {
 
     if ($valid) {
         $aErrormessage = null;
-        // Create a new settings object, and save the settings!
-        $oSettings = new OA_Admin_Settings();
+        // Save the settings!
         $result = $oSettings->processSettingsFromForm($aElements);
         if ($result) {
             // Queue confirmation message